requirements.txt
# Core
pytz==2018.9
Django==2.2.19
django-configurations==2.1
gunicorn==19.9.0
# For the persistence stores
# psycopg2-binary==2.8.3
django-resized==0.3.11
dj-database-url==0.5.0
# Rest apis
djangorestframework==3.9.1
# Markdown==3.0.1
django-filter==2.2.0

# API Documentation
drf-yasg==1.14.0 # 1.16.1 error
# Cors Handling
django-cors-headers==3.0.2
# Image-processing
opencv-python==4.5.1.48
pillow>=6.2.0
django-imagekit==4.0.2
pilkit==2.0
django-imagekit==4.0.2

bitstring==3.1.7

django-debug-toolbar==2.0
django-extensions==2.1.9
Werkzeug==0.15.5
# Task Queue
celery==4.3.0
tornado==5.1.1
flower==0.9.3
kombu==4.4.0
djangorestframework_simplejwt==4.3.0
djangorestframework-jwt==1.11.0
django-rest-auth==0.9.5
django-allauth==0.39.1
python-dateutil==2.8.1
# Celery
# gevent==1.4.0
# Cache
#django-redis==4.10.0
django-cacheops==4.2
django-taggit==1.5.1
six==1.16.0
psycopg2-binary==2.8.6